chore:SP-2340 Adds crypto hints to cryptography table
