+++ src/main/model/entity/Cryptography.ts
@@ -4,6 +4,17 @@ export class Cryptography {
   version: string;
 
   algorithms: Array<Algorithms>;
+
+  hints: Array<Hint>;
+}
+
+export interface Hint {
+  id: string;
+  name: string;
+  description: string;
+  category: string;
+  url: string;
+  purl: string;
 }
 
 export interface Algorithms {
